+++ config.ini
@@ -10,7 +10,7 @@ type = Mock
 
 [grbl_streamer]
 mock = False
-type = Arduino
+# type = Arduino
 baudrate = 115200
 
 [windows]
@@ -72,8 +72,17 @@ grbl_x_axis_config = grbl_rot_axis
 grbl_y_axis_config = grbl_rot_axis # read, but y axi is not used
 
 [scanner]
+controller = dual_controllers
+
+[dual_controllers]
+type = DualControllers
 rotator = ESP32DuinoRotation  # tic
 planar_mover_controller = fluidnc2  # grbl_streamer
+feed_rate = 35000
+
+[combined_controller]
+type = CombinedController
+controller = fluidnc2
 feed_rate = 35000
 
 [sweep]
@@ -94,3 +103,5 @@ nr_of_points = 5000
 wall_spacing = 30  # mm
 radius = 320  # mm
 
+[motion_manager]
+type = CylindricalMeasurementMotionManager
